добавляет контроллер
